Move contents of the Discover component into according components
